requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.13
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astroid==2.15.8
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
bandit==1.7.5
biopython==1.81
black==23.9.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
coverage==7.3.2
coverage-badge==1.1.0
cycler==0.12.1
darglint==1.8.1
dill==0.3.7
distlib==0.3.7
docker-pycreds==0.4.0
docutils==0.18.1
dparse==0.6.3
exceptiongroup==1.1.3
fair-esm==2.0.0
fast-pytorch-kmeans==0.2.0.1
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.37
h5py==3.10.0
h5py-cache==1.0
huggingface-hub==0.17.3
identify==2.5.30
idna==3.4
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lightning-utilities==0.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy==0.910
mypy-extensions==0.4.4
networkx==3.1
nodeenv==1.8.0
numpy==1.25.2
omegaconf==2.3.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.2
pathtools==0.1.2
pbr==5.11.1
Pillow==10.0.1
pip==23.2.1
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
protobuf==4.24.4
psutil==5.9.5
py==1.11.0
pydocstyle==6.3.0
Pygments==2.16.1
pylint==2.17.7
pynvml==11.5.0
pyparsing==3.1.1
PyTDC==0.3.9
pytest==7.4.2
pytest-cov==3.0.0
pytest-html==3.2.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
pytorch-lightning==1.9.5
pytz==2023.3.post1
pyupgrade==3.15.0
PyYAML==6.0.1
rdkit==2022.9.5
rdkit-pypi==2022.9.5
regex==2023.10.3
requests==2.31.0
rich==13.6.0
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safetensors==0.4.0
safety==2.4.0b1
scikit-learn==1.3.1
scipy==1.9.3
seaborn==0.13.0
sentry-sdk==1.32.0
setproctitle==1.3.3
setuptools==68.2.2
setuptools-scm==8.0.4
shellingham==1.5.3
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stevedore==5.1.0
sympy==1.12
threadpoolctl==3.2.0
tokenize-rt==5.2.0
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
torch==1.13.1+cu116
torchmetrics==0.11.4
tqdm==4.66.1
transformers==4.34.0
typer==0.4.2
typing_extensions==4.8.0
urllib3==2.0.6
virtualenv==20.24.5
wandb==0.13.11
wheel==0.41.2
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0